added CSVs
The individual contributions data was too large to upload all at once here (over 25GB), so I split it up. The split is uneven, since the size of different years varies greatly. 2022 is weirdly big. I have the full data from 04 to 23 in one table, just need to find a place to host it, contact me about this.